Add some casts to quiet -Wconversion

+++ src/conversation.c
@@ -219,7 +219,7 @@ sudo_conversation_printf(int msg_type, const char * restrict fmt, ...)
             va_end(ap);
         }
         if (len != -1) {
-            if (fwrite(buf, 1, len, ttyfp ? ttyfp : fp) == 0)
+            if (fwrite(buf, 1, (size_t)len, ttyfp ? ttyfp : fp) == 0)
                 len = -1;
             if (buf != sbuf)
                 free(buf);
